refactor: trendlines
