+++ package.json
@@ -67,7 +67,7 @@
 		"@testing-library/jest-dom": "^5.16.5",
 		"@testing-library/react": "12.1.5",
 		"@testing-library/user-event": "^14.4.3",
-		"@trivago/prettier-plugin-sort-imports": "^4.0.0",
+		"@trivago/prettier-plugin-sort-imports": "^4.3.0",
 		"@types/node": "^20.8.2",
 		"@types/react": "17.0.50",
 		"@types/react-dom": "17.0.17",
