+++ src/specBuilder/trendline/trendlineScaleUtils.test.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2024 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import { createElement } from 'react';
+import { getTrendlineScales } from './trendlineScaleUtils';
+import { defaultLineProps } from './trendlineTestUtils';
+import { Trendline } from '@components/Trendline';
+
+describe('getTrendlineScales()', () => {
+	test('should return the xTrendline scale if the scaleType is time and there is a regression trendline', () => {
+		const scales = getTrendlineScales({
+			...defaultLineProps,
+			children: [createElement(Trendline, { method: 'linear' })],
+		});
+		expect(scales).toHaveLength(1);
+		expect(scales[0]).toHaveProperty('name', 'xTrendline');
+	});
+});
